Move bokeh directory to dashboard
