+++ distributed/dashboard/static/css/base.css
