+++ distributed/dashboard/templates/base.html
