+++ distributed/dashboard/templates/json-index.html
