+++ distributed/dashboard/templates/logs.html
