+++ distributed/dashboard/templates/simple.html
