+++ distributed/dashboard/templates/status.html
