+++ distributed/dashboard/templates/task.html
