+++ distributed/dashboard/templates/worker-table.html
