+++ distributed/dashboard/templates/worker.html
