+++ distributed/dashboard/templates/workers.html
